+++ weaver/wps/__init__.py
@@ -9,7 +9,7 @@ import logging
 from cornice.service import Service
 from pyramid.settings import asbool
 
-from weaver.formats import OUTPUT_FORMAT_XML
+from weaver.formats import CONTENT_TYPE_TEXT_XML, OUTPUT_FORMAT_XML
 from weaver.utils import get_settings
 from weaver.wps.utils import get_wps_path
 
@@ -25,13 +25,13 @@ def includeme(config):
     else:
         logger.debug("Weaver WPS enabled.")
         wps_path = get_wps_path(settings)
-        wps_service = Service(name="wps", path=wps_path)
+        wps_service = Service(name="wps", path=wps_path, content_type=CONTENT_TYPE_TEXT_XML)
         logger.debug("Adding WPS KVP/XML schemas.")
         wps_tags = [sd.TAG_GETCAPABILITIES, sd.TAG_DESCRIBEPROCESS, sd.TAG_EXECUTE, sd.TAG_WPS]
         wps_service.add_view("GET", pywps_view, tags=wps_tags, renderer=OUTPUT_FORMAT_XML,
-                             schema=sd.WPSEndpoint(), response_schemas=sd.wps_responses)
+                             schema=sd.WPSEndpointGet(), response_schemas=sd.wps_responses)
         wps_service.add_view("POST", pywps_view, tags=wps_tags, renderer=OUTPUT_FORMAT_XML,
-                             schema=sd.WPSEndpoint(), response_schemas=sd.wps_responses)
+                             schema=sd.WPSEndpointPost(), response_schemas=sd.wps_responses)
         logger.debug("Adding WPS KVP/XML view.")
         config.add_route(**sd.service_api_route_info(wps_service, settings))
         config.add_view(pywps_view, route_name=wps_service.name)