+++ CHANGES.rst
@@ -16,10 +16,16 @@ Changes:
   (see issue `opengeospatial/ogcapi-processes#124 <https://github.com/opengeospatial/ogcapi-processes/issues/124>`_ and
   PR `opengeospatial/ogcapi-processes#159 <https://github.com/opengeospatial/ogcapi-processes/pull/159>`_, resolves
   `#235 <https://github.com/crim-ca/weaver/issues/235>`_).
+- Add OpenAPI schema examples for some of the most common responses.
+- Add missing schema definitions for WPS XML requests and responses.
+- Improve schema self-validation with their specified default values.
+- Add explicit options usage and expected parsing results for all test variations of OpenAPI schemas generation and
+  ``colander`` object arguments for future reference in ``tests.wps_restapi.test_colander_extras``.
 
 Fixes:
 ------
-- No change.
+- Fix erroneous tags in job inputs schemas.
+- Fix handling of deeply nested schema validator raising for invalid format within optional parent schema.
 
 `3.0.0 <https://github.com/crim-ca/weaver/tree/3.0.0>`_ (2021-03-16)
 ========================================================================